Feature/multi

+++ pkg/kubelet/subscriber/subscriber.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"github.com/IBM/sarama"
 	"minik8s/pkg/api/msg_type"
+	"minik8s/pkg/config"
 	"minik8s/pkg/kafka"
 	"minik8s/pkg/kubelet/host"
 	"minik8s/pkg/kubelet/node"
@@ -21,12 +22,11 @@ type KubeletSubscriber struct {
 }
 
 func NewKubeletSubscriber() *KubeletSubscriber {
-	brokers := []string{"127.0.0.1:9092"}
-	group := "kubelet"
+	group := "kubelet" + "-" + config.Nodename
 	return &KubeletSubscriber{
 		ready:       make(chan bool),
 		done:        make(chan bool),
-		subscriber:  kafka.NewSubscriber(brokers, group),
+		subscriber:  kafka.NewSubscriber(group),
 		HostManager: host.NewHostManager(),
 	}
 }
@@ -64,10 +64,16 @@ func (k *KubeletSubscriber) PodHandler(msg []byte) {
 	}
 	switch podMsg.Opt {
 	case msg_type.Add:
+		if podMsg.NewPod.Spec.NodeName != config.Nodename {
+			break
+		}
 		log.Info("create pod: %v", podMsg.NewPod)
 		pod.CreatePod(&podMsg.NewPod)
 		break
 	case msg_type.Delete:
+		if podMsg.OldPod.Spec.NodeName != config.Nodename {
+			break
+		}
 		log.Info("delete pod: %v", podMsg.OldPod)
 		pod.DeletePod(&podMsg.OldPod)
 		break
@@ -76,8 +82,12 @@ func (k *KubeletSubscriber) PodHandler(msg []byte) {
 		NewSpec, _ := json.Marshal(podMsg.NewPod.Spec)
 		if string(OldSpec) != string(NewSpec) {
 			log.Info("update pod: %v", podMsg.NewPod)
-			pod.DeletePod(&podMsg.OldPod)
-			pod.CreatePod(&podMsg.NewPod)
+			if podMsg.OldPod.Spec.NodeName == config.Nodename {
+				pod.DeletePod(&podMsg.OldPod)
+			}
+			if podMsg.NewPod.Spec.NodeName == config.Nodename {
+				pod.CreatePod(&podMsg.NewPod)
+			}
 		}
 		break
 	}
